refactor(31400): refactor the group creation - fix potential bug with deprecated getGroupLayout - change link style to dynamic - export creation function
